refactor: criado componente DisplayResults para renderizar os resultados
